Extract gzip/binary decoders into internal/serialize
The metadata loader and the prefix lookups (carrier/geocoding/timezone/region)
share these codecs, so they move to a shared internal/serialize package
(LoadPrefixMap/LoadIntArrayMap/DecodeUnzip) rather than being split by concern.
Prerequisite for moving the metadata source into the metadata package.
