+++ shortnumberinfo.go
@@ -6,6 +6,7 @@ import (
 	"slices"
 
 	"github.com/nyaruka/phonenumbers/v2/internal/regexbasedmatcher"
+	"github.com/nyaruka/phonenumbers/v2/internal/serialize"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -39,7 +40,7 @@ func ShortNumberMetadataCollection() (*PhoneMetadataCollection, error) {
 		return currShortNumberMetadataColl, nil
 	}
 
-	rawBytes, err := decodeUnzip(shortNumberData)
+	rawBytes, err := serialize.DecodeUnzip(shortNumberData)
 	if err != nil {
 		return nil, err
 	}
